# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MHED Platform API
  version: 0.2.0
  description: Public plaza, administration, authentication, and MH3G game-data API.
servers:
  - url: /
paths:
...
  /v1/admin/users:
    get:
      operationId: listUsers
      security: [cookieSession: []]
      responses: { "200": { description: User list } }
    post:
      operationId: createUser
      security: [cookieSession: []]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      responses: { "201": { description: User and one-time temporary password }, "409": { $ref: "#/components/responses/Error" } }
  /v1/admin/users/{id}/status:
    patch:
      operationId: updateUserStatus
      security: [cookieSession: []]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/CSRFToken" }]
      responses: { "204": { description: Status updated }, "409": { $ref: "#/components/responses/Error" } }
  /v1/admin/users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      security: [cookieSession: []]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/CSRFToken" }]
      responses: { "200": { description: One-time temporary password } }
  /v1/admin/users/{id}/roles:
    put:
      operationId: replaceUserRoles
      security: [cookieSession: []]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/CSRFToken" }]
      responses: { "204": { description: Role assignments replaced } }
  /v1/admin/roles:
    get:
      operationId: listRoles
      security: [cookieSession: []]
      responses: { "200": { description: Roles and permission keys } }
    post:
      operationId: createRole
      security: [cookieSession: []]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      responses: { "201": { description: Role created } }
  /v1/admin/roles/{id}:
    put:
      operationId: updateRole
      security: [cookieSession: []]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/CSRFToken" }]
      responses: { "204": { description: Role updated } }
    delete:
      operationId: deleteRole
      security: [cookieSession: []]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/CSRFToken" }]
      responses: { "204": { description: Role deleted }, "404": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    cookieSession: { type: apiKey, in: cookie, name: mhed_admin_session }
    desktopBearer: { type: http, scheme: bearer, bearerFormat: opaque }
  parameters:
    ResourceID: { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    CSRFToken: { in: header, name: X-CSRF-Token, required: true, schema: { type: string } }
  responses:
...
    Error:
      description: Stable API error
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }